requirements.txt
django
djangorestframework
python-dotenv
# psycopg2
pandas
gunicorn
gevent
drf-yasg